// src/types/Item.ts
export interface Item {
    id: string;
    name: string;
    price: number;
    image: string; //FILLER I DONT KNOW WHAT TO PUT HERE
    //secondary feature, filter through different collection, will have to have a collection 
    description: string;
    // collection: string; ***commenting this out here because are no longer distinguishing between figurines and plushes collections

  }